Fix total price after purchase not appear in the order card Fix purchased items not appearing in the details table Fix checkout process not completing Fix other minor issues

+++ lib/data/model/items.dart
@@ -120,8 +120,7 @@ class Item {
       favorite: map['favorite'] as int?,
       itemsCategoryName: map['category_name'] as String?,
       itemsCategoryNameAr: map['category_name_ar'] as String?,
-      itemsPriceAfterDiscount:
-          map['items_price_after_discount'].toDouble() as double?,
+      itemsPriceAfterDiscount: map['items_price_after_discount']?.toDouble(),
     );
   }
